+++ backend/src/customers/customers.module.ts
@@ -2,9 +2,10 @@ import { Module } from '@nestjs/common';
 import { CustomersService } from './customers.service';
 import { CustomersController } from './customers.controller';
 import { LogsModule } from '../logs/logs.module';
+import { CryptoModule } from '../common/crypto/crypto.module';
 
 @Module({
-  imports: [LogsModule],
+  imports: [LogsModule, CryptoModule],
   controllers: [CustomersController],
   providers: [CustomersService],
   exports: [CustomersService],